was added a functional form who performs an inser and delete to 1 to many relationships in some model's class

+++ trunk/ ubicanegocios --username kristiandamian/altaajax.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import cgi
-from modelos import idiomas, estados, tipos_lugares
+from modelos import idiomas, estados, tipos_lugares, ciudades
 from google.appengine.ext import db
 from google.appengine.ext import webapp
 
@@ -20,6 +20,12 @@ class altaajax(webapp.RequestHandler):
             instanciaDB=idiomas(idioma=dato)            
         elif objeto=="Estados": 
             instanciaDB=estados(estado=dato)
+        elif objeto=="Ciudades":
+            estado=self.request.get('Estado')
+            query="where estado = '"+estado+"' " 
+            Estado=estados.gql(query )
+            ObjEstado=Estado.fetch(1) #solo obtengo el primer registro, para evitar duplicados
+            instanciaDB=ciudades(parent=ObjEstado[0],ciudad=dato)            
         else:
             instanciaDB=tipos_lugares(tipo=dato)            
         instanciaDB.put()
@@ -37,13 +43,30 @@ class bajaajax(webapp.RequestHandler):
         elif objeto=="Estados": 
             DB="estados"
             campo="estado"
+        elif objeto=="Ciudades":
+            estado = self.request.get('Estado')
+            DB=None
+            campo=None
+            #Borro las ciudades que se llamen igual, pero SOLO de ese estado
+            query="where estado = '"+estado+"' " 
+            Estado=estados.gql(query)
+            ObjEstado=Estado.fetch(1) #solo obtengo el primer registro, para evitar duplicados
+            
+            for lstcd in ObjEstado:
+                cities=ciudades.all().ancestor(lstcd)
+                query="where ciudad = '"+dato+"' " 
+                for city in cities:
+                    if city.ciudad==dato:
+                        city.delete()
         else:
             DB="tipos_lugares"
             campo="tipo"
+        
+        if not DB==None:
+            q = db.GqlQuery("SELECT * FROM "+DB+" WHERE "+campo+" = '"+dato+"'")
+            for campos in q:
+                campos.delete() #ejecuto la consulta y borro uno por uno
             
-        q = db.GqlQuery("SELECT * FROM "+DB+" WHERE "+campo+" = '"+dato+"'")
-        for campos in q:
-            campos.delete() #ejecuto la consulta y borro uno por uno
         return False
     
 class objAjax (webapp.RequestHandler):
@@ -73,6 +96,19 @@ class objAjax (webapp.RequestHandler):
             jsondata+=[self.addKey(jsondic,"Dato",datos.tipo)]
         return jsondata
     
+    def obtengoCiudades(self, estado):
+        jsondic={}
+        jsondata=[]
+        query="where estado = '"+estado+"' " 
+        Estado=estados.gql(query)
+        ObjEstado=Estado.fetch(1) #solo obtengo el primer registro, para evitar duplicados
+        
+        for lstcd in ObjEstado:
+            cities=ciudades.all().ancestor(lstcd)
+            for city in cities:
+                jsondata+=[self.addKey(jsondic,"Dato",city.ciudad)]
+        return jsondata
+    
     def post(self):
         objeto = self.request.get('tipoObjeto')        
         jsondata=[]
@@ -80,9 +116,9 @@ class objAjax (webapp.RequestHandler):
             jsondata=self.obtengoidiomas()
         elif objeto=="Estados": 
             jsondata=self.obtengoestados()
+        elif objeto=="Ciudades":
+            jsondata=self.obtengoCiudades(estado=self.request.get('Estado'))
         else:
             jsondata=self.obtengolugares()                    
         self.response.out.write(simplejson.dumps(jsondata))
         return False
-    
-
